chore: move BIR files under BIR directory
