+++ bir/include/belalang/BIR/IR/BIRDialect.h
@@ -6,12 +6,12 @@
 #include "mlir/IR/Dialect.h"
 #include "mlir/IR/OpImplementation.h"
 
-#include "belalang/IR/BIRDialect.h.inc"
+#include "belalang/BIR/IR/BIRDialect.h.inc"
 
 #define GET_OP_CLASSES
-#include "belalang/IR/BIROps.h.inc"
+#include "belalang/BIR/IR/BIROps.h.inc"
 
 #define GET_TYPEDEF_CLASSES
-#include "belalang/IR/BIRTypes.h.inc"
+#include "belalang/BIR/IR/BIRTypes.h.inc"
 
 #endif // BELALANG_IR_DIALECT_H_
