+++ bir/include/belalang/BIR/IR/BIRTypes.h
@@ -4,6 +4,6 @@
 #include "mlir/IR/Types.h"
 
 #define GET_TYPEDEF_CLASSES
-#include "belalang/IR/BIRTypes.h.inc"
+#include "belalang/BIR/IR/BIRTypes.h.inc"
 
 #endif // BELALANG_IR_TYPES_H_
