+++ bir/lib/IR/BIRDialect.cpp
@@ -1,6 +1,6 @@
-#include "belalang/IR/BIRDialect.h"
+#include "belalang/BIR/IR/BIRDialect.h"
 
-#include "belalang/IR/BIRDialect.cpp.inc"
+#include "belalang/BIR/IR/BIRDialect.cpp.inc"
 
 namespace belalang {
 namespace bir {
@@ -8,12 +8,12 @@ namespace bir {
 void BIRDialect::initialize() {
   addTypes<
 #define GET_TYPEDEF_LIST
-#include "belalang/IR/BIRTypes.cpp.inc"
+#include "belalang/BIR/IR/BIRTypes.cpp.inc"
       >();
 
   addOperations<
 #define GET_OP_LIST
-#include "belalang/IR/BIROps.cpp.inc"
+#include "belalang/BIR/IR/BIROps.cpp.inc"
       >();
 }
