+++ bir/lib/Transforms/Constantize.cpp
@@ -1,5 +1,5 @@
-#include "belalang/IR/BIRDialect.h"
-#include "belalang/Passes.h"
+#include "belalang/BIR/IR/BIRDialect.h"
+#include "belalang/BIR/Passes.h"
 #include "mlir/Dialect/Arith/IR/Arith.h"
 #include "mlir/IR/BuiltinOps.h"
 #include "mlir/IR/PatternMatch.h"
