+++ bir/src/lib.rs
@@ -3,7 +3,7 @@ pub use ffi::*;
 #[cxx::bridge(namespace = "belalang::bir")]
 mod ffi {
     unsafe extern "C++" {
-        include!("belalang/IR/Builder.h");
+        include!("belalang/BIR/IR/Builder.h");
 
         type BIRValue;
         type BIRBuilder;
